Switch to pages

+++ naturinorge_guide/lib/pages/nin_structure/major_type/minor_type_table/table_body.dart
@@ -23,7 +23,7 @@ class TableBody extends StatelessWidget {
         .toList();
     var staggeredTiles = Provider.of<MajorTypeProvider>(context)
         .minorTypeScaledBlocks
-        .map((e) => StaggeredTile.count(e.width, e.height))
+        .map((e) => StaggeredTile.count(e.width, e.height.toDouble()))
         .toList();
     return Stack(
       children: [
